+++ CHANGES.md
@@ -8,6 +8,11 @@ Known issues:
 
 (nothing yet)
 
+## 5.5.0
+
+- [PUBAPI-1452] Add ip subcommand to network, e.g.
+  `triton network ip`.
+
 ## 5.4.0
 
 - [joyent/node-triton#74, TOOLS-1872] Filter instance list by tag, e.g.